remoção de método e implementação na classe computer repository
