+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database; 
-using LabManeger.Repositories; 
+using LabManeger.Repositories;
+using LabManager.Models;
 
 var databaseConfig = new DatabaseConfig();
 
@@ -30,18 +31,8 @@ if (modelName == "Computer")
         var ram = args[3];
         var processor = args [4]; 
         
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
-
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)"; 
-        command.Parameters.AddWithValue("$id",id);
-        command.Parameters.AddWithValue("$ram",ram);
-        command.Parameters.AddWithValue("processor",processor);
-
-            
-        command.ExecuteNonQuery();
-        connection.Close();
+        var computer = new Computer(id, ram, processor); 
+        computerRepository.Save(computer); 
 
     }
 }
